Split theme out of utils
