+++ src/components/explore/explore-menu.js
@@ -4,7 +4,7 @@ import styled from "styled-components"
 
 import SortMenu from "./sort-menu"
 import { NEGATIVE } from "../../utils/constants"
-import { colors } from "../../utils/theme"
+import { colors } from "../../theme"
 
 const TabButton = styled.button`
   user-select: none;
