+++ src/components/external-link.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types"
 
 import { POSITIVE, NEGATIVE } from "../utils/constants"
 import { PropTypeIsUrl } from "../utils/helpers"
-import { colors } from "../utils/theme"
+import { colors } from "../theme"
 
 export default function ExternalLink({ label, url, id, mode = NEGATIVE }) {
   return (
