+++ src/components/layout/definition-list.js
@@ -3,7 +3,7 @@ import styled from "styled-components"
 import PropTypes from "prop-types"
 
 import { POSITIVE, NEGATIVE } from "../../utils/constants"
-import { colors } from "../../utils/theme"
+import { colors } from "../../theme"
 
 const List = styled.dl`
   display: grid;
