+++ src/components/map/mapbox-gl-draw-styles.js
@@ -1,5 +1,5 @@
 import { NEGATIVE } from "../../utils/constants"
-import { colors } from "../../utils/theme"
+import { colors } from "../../theme"
 
 export const drawStyles = [
   {
