+++ src/templates/focus/index.js
@@ -15,7 +15,7 @@ import ExternalLink from "../../components/external-link"
 import RelatedCampaignsSection from "../../components/related-campaigns-section"
 import FocusAreaGallery from "../../components/focus-area-gallery"
 import { NEGATIVE } from "../../utils/constants"
-import { colors } from "../../utils/theme"
+import { colors } from "../../theme"
 import { getFocusIcon } from "../../icons/utils"
 
 const FocusTemplate = ({ data }) => {
