--- src/utils/README.md
@@ -1,45 +1,0 @@
-# Theme Application
-
-## Colors:
-
-### Font
-
-- main: base
-- list or nav options
-  - active text: base
-  - inactive: grey
-- link: link (color)
-
-### Components:
-
-background: primary
-box background: secondary
-
-## Typography
-
-### Page:
-
-- title: H1
-- subtitle
-- tagline
-
-### Page Header
-
-- title: Bold large text
-- navigation options: normal text
-
-### Page Section
-
-- title: H2
-- tagline
-- content text
-- content title: H3
-
-### Images:
-
-- image title: H3
-
-### labels
-
-- small font uppercase
-- included: icon label, page content sub-group label, chip text, timeline labels, deployment box
